Adjusting & consolidated cron views for new UI.  Created cron dashboard at /cron
